restructure module for win width
